Scheme form is done jsut needs to be styled. Need update actions etc on controller

+++ db/migrate/20140722174854_create_grades.rb
@@ -2,7 +2,7 @@ class CreateGrades < ActiveRecord::Migration
   def change
     create_table :grades do |t|
       t.string :title
-      t.decimal :final_score
+      t.integer :final_score
       t.integer :category_id
       t.integer :student_id
       t.integer :eval_id
