+++ db/migrate/20140727164259_create_scores.rb
@@ -1,7 +1,7 @@
 class CreateScores < ActiveRecord::Migration
   def change
     create_table :scores do |t|
-      t.decimal :score
+      t.integer :score
       t.integer :grade_id
 
       t.timestamps
